Refactoring and test fixes
