+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/functional/TestColumnStatsIndex.scala
@@ -19,7 +19,7 @@
 package org.apache.hudi.functional
 
 import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.{LocatedFileStatus, Path}
+import org.apache.hadoop.fs.Path
 import org.apache.hudi.ColumnStatsIndexSupport.composeIndexSchema
 import org.apache.hudi.DataSourceWriteOptions.{PRECOMBINE_FIELD, RECORDKEY_FIELD}
 import org.apache.hudi.HoodieConversionUtils.toProperties
@@ -28,57 +28,22 @@ import org.apache.hudi.common.model.HoodieTableType
 import org.apache.hudi.common.table.{HoodieTableConfig, HoodieTableMetaClient}
 import org.apache.hudi.common.util.ParquetUtils
 import org.apache.hudi.config.HoodieWriteConfig
-import org.apache.hudi.functional.TestColumnStatsIndex.ColumnStatsTestCase
-import org.apache.hudi.testutils.HoodieSparkClientTestBase
-import org.apache.hudi.{ColumnStatsIndexSupport, DataSourceReadOptions, DataSourceWriteOptions}
+import org.apache.hudi.functional.ColumnStatIndexTestBase.ColumnStatsTestCase
+import org.apache.hudi.{ColumnStatsIndexSupport, DataSourceWriteOptions}
 import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.analysis.UnresolvedAttribute
 import org.apache.spark.sql.catalyst.expressions.{And, AttributeReference, GreaterThan, Literal, Or}
-import org.apache.spark.sql.functions.typedLit
 import org.apache.spark.sql.hudi.DataSkippingUtils.translateIntoColumnStatsIndexFilterExpr
 import org.apache.spark.sql.types._
 import org.junit.jupiter.api.Assertions.{assertEquals, assertNotNull, assertTrue}
 import org.junit.jupiter.api._
 import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.{Arguments, EnumSource, MethodSource, ValueSource}
+import org.junit.jupiter.params.provider.{EnumSource, MethodSource, ValueSource}
 
-import java.math.BigInteger
-import java.sql.{Date, Timestamp}
 import scala.collection.JavaConverters._
-import scala.util.Random
 
 @Tag("functional")
-class TestColumnStatsIndex extends HoodieSparkClientTestBase {
-  var spark: SparkSession = _
-
-  val sourceTableSchema =
-    new StructType()
-      .add("c1", IntegerType)
-      .add("c2", StringType)
-      .add("c3", DecimalType(9, 3))
-      .add("c4", TimestampType)
-      .add("c5", ShortType)
-      .add("c6", DateType)
-      .add("c7", BinaryType)
-      .add("c8", ByteType)
-
-  @BeforeEach
-  override def setUp() {
-    initPath()
-    initSparkContexts()
-    initFileSystem()
-
-    setTableName("hoodie_test")
-    initMetaClient()
-
-    spark = sqlContext.sparkSession
-  }
-
-  @AfterEach
-  override def tearDown() = {
-    cleanupFileSystem()
-    cleanupSparkContexts()
-  }
+class TestColumnStatsIndex extends ColumnStatIndexTestBase {
 
   @ParameterizedTest
   @MethodSource(Array("testMetadataColumnStatsIndexParams"))
@@ -125,89 +90,6 @@ class TestColumnStatsIndex extends HoodieSparkClientTestBase {
       saveMode = SaveMode.Append)
   }
 
-  @ParameterizedTest
-  @MethodSource(Array("testMetadataColumnStatsIndexParams"))
-  def testMetadataColumnStatsIndexWithSQL(testCase: ColumnStatsTestCase): Unit = {
-    val metadataOpts = Map(
-      HoodieMetadataConfig.ENABLE.key -> "true",
-      HoodieMetadataConfig.ENABLE_METADATA_INDEX_COLUMN_STATS.key -> "true"
-    )
-
-    var commonOpts = Map(
-      "hoodie.insert.shuffle.parallelism" -> "4",
-      "hoodie.upsert.shuffle.parallelism" -> "4",
-      HoodieWriteConfig.TBL_NAME.key -> "hoodie_test",
-      DataSourceWriteOptions.TABLE_TYPE.key -> testCase.tableType.toString,
-      RECORDKEY_FIELD.key -> "c1",
-      PRECOMBINE_FIELD.key -> "c1",
-      HoodieTableConfig.POPULATE_META_FIELDS.key -> "true",
-      DataSourceReadOptions.ENABLE_DATA_SKIPPING.key -> "true",
-      DataSourceReadOptions.QUERY_TYPE.key -> DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL
-    ) ++ metadataOpts
-
-    doWriteAndValidateColumnStats(testCase, metadataOpts, commonOpts,
-      dataSourcePath = "index/colstats/input-table-json",
-      expectedColStatsSourcePath = "index/colstats/column-stats-index-table.json",
-      operation = DataSourceWriteOptions.INSERT_OPERATION_OPT_VAL,
-      saveMode = SaveMode.Overwrite)
-
-    doWriteAndValidateColumnStats(testCase, metadataOpts, commonOpts,
-      dataSourcePath = "index/colstats/another-input-table-json",
-      expectedColStatsSourcePath = "index/colstats/updated-column-stats-index-table.json",
-      operation = DataSourceWriteOptions.INSERT_OPERATION_OPT_VAL,
-      saveMode = SaveMode.Append)
-
-    // NOTE: MOR and COW have different fixtures since MOR is bearing delta-log files (holding
-    //       deferred updates), diverging from COW
-    val expectedColStatsSourcePath = if (testCase.tableType == HoodieTableType.COPY_ON_WRITE) {
-      "index/colstats/cow-updated2-column-stats-index-table.json"
-    } else {
-      "index/colstats/mor-updated2-column-stats-index-table.json"
-    }
-
-    createSQLTable(commonOpts, DataSourceReadOptions.QUERY_TYPE_SNAPSHOT_OPT_VAL)
-    val c5GT50WithDataSkipping = spark.sql("select * from tbl where c5 > 70").count()
-
-    doWriteAndValidateColumnStats(testCase, metadataOpts, commonOpts,
-      dataSourcePath = "index/colstats/update-input-table-json",
-      expectedColStatsSourcePath = expectedColStatsSourcePath,
-      operation = DataSourceWriteOptions.UPSERT_OPERATION_OPT_VAL,
-      saveMode = SaveMode.Append)
-
-    // verify snapshot query
-    verifySQLQueries(c5GT50WithDataSkipping, DataSourceReadOptions.QUERY_TYPE_SNAPSHOT_OPT_VAL, commonOpts)
-
-    // verify read_optimized query
-    verifySQLQueries(c5GT50WithDataSkipping, DataSourceReadOptions.QUERY_TYPE_READ_OPTIMIZED_OPT_VAL, commonOpts)
-
-    // verify incremental query
-    verifySQLQueries(c5GT50WithDataSkipping, DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL, commonOpts)
-    commonOpts = commonOpts + (DataSourceReadOptions.INCREMENTAL_FALLBACK_TO_FULL_TABLE_SCAN_FOR_NON_EXISTING_FILES.key -> "true")
-    verifySQLQueries(c5GT50WithDataSkipping, DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL, commonOpts)
-  }
-
-  private def verifySQLQueries(c5GT50WithDataSkippingAtPrevInstant: Long, queryType: String, opts: Map[String, String]): Unit = {
-    // 2 records are updated with c5 greater than 70 and one record is inserted with c5 value greater than 70
-    var commonOpts:Map[String, String] = opts
-    createSQLTable(commonOpts, queryType)
-    val increment = if (queryType.equals(DataSourceReadOptions.QUERY_TYPE_READ_OPTIMIZED_OPT_VAL) && metaClient.getTableType == HoodieTableType.MERGE_ON_READ) {
-      1 // only one insert
-    } else {
-      3 // one insert and two upserts
-    }
-    assertEquals(spark.sql("select * from tbl where c5 > 70").count(), c5GT50WithDataSkippingAtPrevInstant + increment)
-    val c5GT50WithDataSkipping = spark.sql("select * from tbl where c5 > 70").count()
-
-    if (queryType.equals(DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL)) {
-      createIncrementalSQLTable(commonOpts, metaClient.reloadActiveTimeline().getInstants.get(1).getTimestamp)
-      assertEquals(spark.sql("select * from tbl where c5 > 70").count(), 3)
-    }
-
-    commonOpts = opts + (DataSourceReadOptions.ENABLE_DATA_SKIPPING.key -> "false")
-    createSQLTable(commonOpts, queryType)
-    val c5Gt50WithoutDataSkipping = spark.sql("select * from tbl where c5 > 70").count()
-    assertEquals(c5GT50WithDataSkipping, c5Gt50WithoutDataSkipping)
-  }
 
   @ParameterizedTest
   @EnumSource(classOf[HoodieTableType])
@@ -537,220 +419,5 @@ class TestColumnStatsIndex extends HoodieSparkClientTestBase {
       assertNotNull(max)
       assertTrue(r.getMinValue.asInstanceOf[Comparable[Object]].compareTo(r.getMaxValue.asInstanceOf[Object]) <= 0)
     })
-  }
-
-  private def createSQLTable(hudiOpts: Map[String, String], queryType: String): Unit = {
-    val opts = hudiOpts + (
-      DataSourceReadOptions.QUERY_TYPE.key -> queryType,
-      DataSourceReadOptions.BEGIN_INSTANTTIME.key() -> metaClient.getActiveTimeline.getInstants.get(0).getTimestamp.replaceFirst(".", "0")
-    )
-    val inputDF1 = spark.read.format("hudi").options(opts).load(basePath)
-    inputDF1.createOrReplaceTempView("tbl")
-  }
-
-  private def createIncrementalSQLTable(hudiOpts: Map[String, String], instantTime: String): Unit = {
-    val opts = hudiOpts + (
-      DataSourceReadOptions.QUERY_TYPE.key -> DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL,
-      DataSourceReadOptions.BEGIN_INSTANTTIME.key() -> instantTime
-    )
-    val inputDF1 = spark.read.format("hudi").options(opts).load(basePath)
-    inputDF1.createOrReplaceTempView("tbl")
-  }
-
-  private def doWriteAndValidateColumnStats(testCase: ColumnStatsTestCase,
-                                            metadataOpts: Map[String, String],
-                                            hudiOpts: Map[String, String],
-                                            dataSourcePath: String,
-                                            expectedColStatsSourcePath: String,
-                                            operation: String,
-                                            saveMode: SaveMode): Unit = {
-    val sourceJSONTablePath = getClass.getClassLoader.getResource(dataSourcePath).toString
-
-    // NOTE: Schema here is provided for validation that the input date is in the appropriate format
-    val inputDF = spark.read.schema(sourceTableSchema).json(sourceJSONTablePath)
-
-    inputDF
-      .sort("c1")
-      .repartition(4, new Column("c1"))
-      .write
-      .format("hudi")
-      .options(hudiOpts)
-      .option(HoodieStorageConfig.PARQUET_MAX_FILE_SIZE.key, 10 * 1024)
-      .option(DataSourceWriteOptions.OPERATION.key, operation)
-      .mode(saveMode)
-      .save(basePath)
-
-    metaClient = HoodieTableMetaClient.reload(metaClient)
-
-    // Currently, routine manually validating the column stats (by actually reading every column of every file)
-    // only supports parquet files. Therefore we skip such validation when delta-log files are present, and only
-    // validate in following cases: (1) COW: all operations; (2) MOR: insert only.
-    val shouldValidateColumnStatsManually = testCase.tableType == HoodieTableType.COPY_ON_WRITE ||
-      operation.equals(DataSourceWriteOptions.INSERT_OPERATION_OPT_VAL)
-
-    validateColumnStatsIndex(
-      testCase, metadataOpts, expectedColStatsSourcePath, shouldValidateColumnStatsManually)
-  }
-
-  private def buildColumnStatsTableManually(tablePath: String,
-                                            includedCols: Seq[String],
-                                            indexedCols: Seq[String],
-                                            indexSchema: StructType): DataFrame = {
-    val files = {
-      val it = fs.listFiles(new Path(tablePath), true)
-      var seq = Seq[LocatedFileStatus]()
-      while (it.hasNext) {
-        seq = seq :+ it.next()
-      }
-      seq.filter(fs => fs.getPath.getName.endsWith(".parquet"))
-    }
-
-    spark.createDataFrame(
-      files.flatMap(file => {
-        val df = spark.read.schema(sourceTableSchema).parquet(file.getPath.toString)
-        val exprs: Seq[String] =
-          s"'${typedLit(file.getPath.getName)}' AS file" +:
-          s"sum(1) AS valueCount" +:
-            df.columns
-              .filter(col => includedCols.contains(col))
-              .filter(col => indexedCols.contains(col))
-              .flatMap(col => {
-                val minColName = s"${col}_minValue"
-                val maxColName = s"${col}_maxValue"
-                if (indexedCols.contains(col)) {
-                  Seq(
-                    s"min($col) AS $minColName",
-                    s"max($col) AS $maxColName",
-                    s"sum(cast(isnull($col) AS long)) AS ${col}_nullCount"
-                  )
-                } else {
-                  Seq(
-                    s"null AS $minColName",
-                    s"null AS $maxColName",
-                    s"null AS ${col}_nullCount"
-                  )
-                }
-              })
-
-        df.selectExpr(exprs: _*)
-          .collect()
-      }).asJava,
-      indexSchema
-    )
-  }
-
-  private def validateColumnStatsIndex(testCase: ColumnStatsTestCase,
-                                       metadataOpts: Map[String, String],
-                                       expectedColStatsSourcePath: String,
-                                       validateColumnStatsManually: Boolean): Unit = {
-    val metadataConfig = HoodieMetadataConfig.newBuilder()
-      .fromProperties(toProperties(metadataOpts))
-      .build()
-
-    val columnStatsIndex = new ColumnStatsIndexSupport(spark, sourceTableSchema, metadataConfig, metaClient)
-
-    val indexedColumns: Set[String] = {
-      val customIndexedColumns = metadataConfig.getColumnsEnabledForColumnStatsIndex
-      if (customIndexedColumns.isEmpty) {
-        sourceTableSchema.fieldNames.toSet
-      } else {
-        customIndexedColumns.asScala.toSet
-      }
-    }
-    val (expectedColStatsSchema, _) = composeIndexSchema(sourceTableSchema.fieldNames, indexedColumns, sourceTableSchema)
-    val validationSortColumns = Seq("c1_maxValue", "c1_minValue", "c2_maxValue", "c2_minValue")
-
-    columnStatsIndex.loadTransposed(sourceTableSchema.fieldNames, testCase.shouldReadInMemory) { transposedColStatsDF =>
-      // Match against expected column stats table
-      val expectedColStatsIndexTableDf =
-        spark.read
-          .schema(expectedColStatsSchema)
-          .json(getClass.getClassLoader.getResource(expectedColStatsSourcePath).toString)
-
-      assertEquals(expectedColStatsIndexTableDf.schema, transposedColStatsDF.schema)
-      // NOTE: We have to drop the `fileName` column as it contains semi-random components
-      //       that we can't control in this test. Nevertheless, since we manually verify composition of the
-      //       ColStats Index by reading Parquet footers from individual Parquet files, this is not an issue
-      assertEquals(asJson(sort(expectedColStatsIndexTableDf, validationSortColumns)),
-        asJson(sort(transposedColStatsDF.drop("fileName"), validationSortColumns)))
-
-      if (validateColumnStatsManually) {
-        // TODO(HUDI-4557): support validation of column stats of avro log files
-        // Collect Column Stats manually (reading individual Parquet files)
-        val manualColStatsTableDF =
-          buildColumnStatsTableManually(basePath, sourceTableSchema.fieldNames, sourceTableSchema.fieldNames, expectedColStatsSchema)
-
-        assertEquals(asJson(sort(manualColStatsTableDF, validationSortColumns)),
-          asJson(sort(transposedColStatsDF, validationSortColumns)))
-      }
-    }
-  }
-
-  private def generateRandomDataFrame(spark: SparkSession): DataFrame = {
-    val sourceTableSchema =
-      new StructType()
-        .add("c1", IntegerType)
-        .add("c2", StringType)
-        // NOTE: We're testing different values for precision of the decimal to make sure
-        //       we execute paths bearing different underlying representations in Parquet
-        // REF: https://github.com/apache/parquet-format/blob/master/LogicalTypes.md#DECIMAL
-        .add("c3a", DecimalType(9, 3))
-        .add("c3b", DecimalType(10, 3))
-        .add("c3c", DecimalType(20, 3))
-        .add("c4", TimestampType)
-        .add("c5", ShortType)
-        .add("c6", DateType)
-        .add("c7", BinaryType)
-        .add("c8", ByteType)
-
-    val rdd = spark.sparkContext.parallelize(0 to 1000, 1).map { item =>
-      val c1 = Integer.valueOf(item)
-      val c2 = Random.nextString(10)
-      val c3a = java.math.BigDecimal.valueOf(Random.nextInt() % (1 << 24), 3)
-      val c3b = java.math.BigDecimal.valueOf(Random.nextLong() % (1L << 32), 3)
-      // NOTE: We cap it at 2^64 to make sure we're not exceeding target decimal's range
-      val c3c = new java.math.BigDecimal(new BigInteger(64, new java.util.Random()), 3)
-      val c4 = new Timestamp(System.currentTimeMillis())
-      val c5 = java.lang.Short.valueOf(s"${(item + 16) / 10}")
-      val c6 = Date.valueOf(s"${2020}-${item % 11 + 1}-${item % 28 + 1}")
-      val c7 = Array(item).map(_.toByte)
-      val c8 = java.lang.Byte.valueOf("9")
-
-      RowFactory.create(c1, c2, c3a, c3b, c3c, c4, c5, c6, c7, c8)
-    }
-
-    spark.createDataFrame(rdd, sourceTableSchema)
-  }
-
-  private def asJson(df: DataFrame) =
-    df.toJSON
-      .select("value")
-      .collect()
-      .toSeq
-      .map(_.getString(0))
-      .mkString("\n")
-
-  private def sort(df: DataFrame): DataFrame = {
-    sort(df, Seq("c1_maxValue", "c1_minValue"))
-  }
-
-  private def sort(df: DataFrame, sortColumns: Seq[String]): DataFrame = {
-    val sortedCols = df.columns.sorted
-    // Sort dataset by specified columns (to minimize non-determinism in case multiple files have the same
-    // value of the first column)
-    df.select(sortedCols.head, sortedCols.tail: _*)
-      .sort(sortColumns.head, sortColumns.tail: _*)
-  }
-}
-
-object TestColumnStatsIndex {
-
-  case class ColumnStatsTestCase(tableType: HoodieTableType, shouldReadInMemory: Boolean)
-
-  def testMetadataColumnStatsIndexParams: java.util.stream.Stream[Arguments] = {
-    java.util.stream.Stream.of(HoodieTableType.values().toStream.flatMap(tableType =>
-      Seq(Arguments.arguments(ColumnStatsTestCase(tableType, shouldReadInMemory = true)),
-        Arguments.arguments(ColumnStatsTestCase(tableType, shouldReadInMemory = false)))
-    ): _*)
   }
 }
