+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/spark/sql/hudi/procedure/TestClusteringProcedure.scala
@@ -587,7 +587,7 @@ class TestClusteringProcedure extends HoodieSparkProcedureTestBase {
 
       metaClient.reloadActiveTimeline()
       val fileIndex1 = HoodieFileIndex(spark, metaClient, None, queryOpts)
-      val orderAllFiles = fileIndex1.allFiles.size
+      val orderAllFiles = fileIndex1.allBaseFiles.size
       val c2OrderFilterCount = fileIndex1.listFiles(Seq(), Seq(dataFilterC2)).head.files.size
       val c3OrderFilterCount = fileIndex1.listFiles(Seq(), Seq(dataFilterC3)).head.files.size
 
@@ -604,7 +604,7 @@ class TestClusteringProcedure extends HoodieSparkProcedureTestBase {
 
       metaClient.reloadActiveTimeline()
       val fileIndex2 = HoodieFileIndex(spark, metaClient, None, queryOpts)
-      val ZOrderAllFiles = fileIndex2.allFiles.size
+      val ZOrderAllFiles = fileIndex2.allBaseFiles.size
       val c2ZOrderFilterCount = fileIndex2.listFiles(Seq(), Seq(dataFilterC2)).head.files.size
       val c3ZOrderFilterCount = fileIndex2.listFiles(Seq(), Seq(dataFilterC3)).head.files.size
